+++ src/components/MissionCard.js
@@ -1,0 +1,6 @@
+export function MissionCard({ wikipedia, mission_name, description }) {
+  return `<div>
+            <a href="${wikipedia}">${mission_name}</a>
+            <p>${description}</p>
+        </div>`;
+}
