+++ src/data/dataStore.js
@@ -1,0 +1,9 @@
+export const dataStore = {
+  rockets: [],
+  selectedRocket: 'Falcon 1',
+  missions: [],
+  selectedMission: 'Thaicom',
+  histories: [],
+  isDataLoading: false,
+  error: null,
+};
