+++ src/index.html
@@ -7,6 +7,6 @@
   </head>
   <body>
     <div id="app-root"></div>
-    <script src="./index.js"></script>
+    <script src="index.js"></script>
   </body>
 </html>
